+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Game from './components/game';
+import Game from './components/game/game';
 
 
 const root = ReactDOM.createRoot(
